# src/main/kotlin/ru/itmo/soa/domain/Enums.kt
package ru.itmo.soa.domain

enum class FormOfEducation {
    DISTANCE_EDUCATION,
    FULL_TIME_EDUCATION,
    EVENING_CLASSES,
}

enum class Semester {
    SECOND,
    THIRD,
    SEVENTH,
    EIGHTH,
}

enum class Color {
    GREEN,
    YELLOW,
    ORANGE,
}

enum class Country {
    USA,
    CHINA,
    VATICAN,
    SOUTH_KOREA,
    NORTH_KOREA,
}